[ENG-3436] chore: Make some code refactoring for RBF in the extension (#712)
* [ENG-3436] chore: Make some code refactoring for RBF in the extension

* Make some code fixes

* Make some code fixes

* fix: put speed up screen in another scrollable container

* Make some code tweaks according to PR review comments

* Remove the unused import

---------
