+++ src/locales/en.json
@@ -814,7 +814,12 @@
     "INSUFFICIENT_FEE": "Insufficient fee",
     "MANUAL_SETTING": "Manual setting",
     "TX_FEE_UPDATED": "Transaction fee updated",
-    "FEE_TOO_LOW": "The minimum fee is {{minimumFee}}"
+    "FEE_TOO_LOW": "The minimum fee is {{minimumFee}}",
+    "TIME": {
+      "SEVERAL_HOURS_OR_MORE": "several hours or more",
+      "HOUR": "hour",
+      "MINUTES": "mins"
+    }
   },
   "POST_CONDITION_MESSAGE": {
     "YOU": "You",
